+++ zanata-model/src/main/java/org/zanata/annotation/EntityRestrict.java
@@ -26,7 +26,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import org.zanata.EntityAction;
+import org.zanata.security.EntityAction;
 
 
 /**
