+++ zanata-model/src/main/java/org/zanata/security/EntityAction.java
@@ -1,4 +1,4 @@
-package org.zanata;
+package org.zanata.security;
 
 /**
  * @author Patrick Huang <a
